working out line
trying the method of simply adjusting two y points as a way of stretching the line to fit the space. would need more lines drawn for different distances to properly fill the space, starts the break down after the window is 600px tall, or about 100px taller than the image was drawn to cover.