about button position
is now relative to the window, and the other logo-components that it is meant to line up with.

line is also redrawn on scroll, since iOS will resize the browser without firing a window.resize event. registering the line recalculation on touchend doesn't work either, because the browser chrome is still in flux. scroll is the only event where things have settled into place and can be recalculated.