fixed transition
in moving to a singular svg for the line, that meant transitioning both the about, and the original content in the page. the original page content then had two conflicting transitions on it. the first, was the initial text fade in transition. the second, was the transition to move out of the way of the about overlay. both transitions defined transitioning the transform attribute, with different rules.

the fix was to have the content element have only a class for the intial transition and transform timings, and once that transition ends, the element gets a class that will have it transitioning out of the way of the about overlay.