firefox svg fix
instead of relying on firefox to calculate the original drawn svg delta x and y, they are part of the codebase.

a new line now needs: a scaling function to reference, the svg path instructions for the drawn path (path.attr('d')), the change in x and y between the start and end point of that original drawn path.