scroll performance
added mordernizr to test for touch
if touch, no need to recalculate the line everytime you scroll. that is in place for mobile browsers where the resizing viewport will require a redraw, but there is no event to attach to for that.
saved a reference to the container node in bottom.js, so we dont always have to get that in order to getBoundingClientRect