behance api reorg
instead of calling for every students list of projects, then looking for their details, we are going to try using the project search api endpoint, filtering by our tag, and common mutations. Then we will compare that to the list of students whose user names we have, and only include those that are on our list. This way we save on calling every single project in order to figure out which are tagged with the tag of interest