hash refactor
hash is now set up for the Go! page, in addition to the work. This is a necessary step if we are going to implement a back button. we wouldn't want inconsistent stuff going on. Say you look at some work. Then look at Go!, and you used the back button to get out of the work, and expect the same to happen when you hit Go!. This is now possible