work sizing
had the wrong set of rules for sizing the work's margin such that the height of the grid was enough to fill the window, ensuring that the filters could be top aligned.
also, since the filters are in their own fixed container, the work does not need to be padded so it can be seen. that was causing a large gap to appear between the filters and the work