mobile
mobile department filtering was showing up at the bottom of the window when you scrolled down to the end of a page. upped the top value to be 200%, where previously at 100%, and the changing viewport size, you would see it every once in a while
another side effect was the body was being clipped of its color, causing another strange to transition. to fix it, the html element has the classes that control size and color, instead of them being trapped on the body which wasn't keeping appropriate height
another problem with the viewport changing size is that the line is drawn based on the viewport height. the height would change, the line wouldn't. instead of just having the line recalculate on touchmove (which is the only place to check to see if the height has changed), i added a check to that event. now touch move fires a check on the height. if the height of the window has changed, then recalculate the line. if not, do nothing.